fix test and move all fixtures into conftest.py
